+++ routes/404.js
@@ -1,0 +1,5 @@
+module.exports = function (app, db) {
+  app.get('*', (req, res) => {
+    res.render('404')
+  })
+}
