+++ routes/rankings.js
@@ -1,0 +1,149 @@
+module.exports = function (app, db) {
+  // Rankings Landing Page
+  app.get('/:league_short_name/rankings', (req, res) => {
+    db.one('SELECT id, name from league WHERE short_name=$1', [req.params.league_short_name])
+    .then((league) => {
+      return Promise.all([
+        league,
+        db.manyOrNone(`
+          SELECT * FROM (
+            SELECT
+              player.id,
+              name,
+              color,
+              (SELECT count(id) FROM game WHERE winner_id=player.id AND league_id=$1) as games_won,
+              (SELECT count(id) FROM game WHERE loser_id=player.id AND league_id=$1) as games_lost,
+              (SELECT count(id) FROM game WHERE (winner_id = player.id OR loser_id=player.id) AND league_id=$1) as total_games,
+              player_to_league.elo_rating as elo_rating
+            FROM player
+            INNER JOIN player_to_league ON player_to_league.player_id=player.id
+            WHERE player_to_league.league_id=$1 AND player.is_active=true
+            GROUP BY (player.id, player_to_league.elo_rating)
+          ) as player_stats
+          WHERE player_stats.total_games > 0
+          ORDER BY elo_rating DESC
+        `, [league.id]),
+        db.manyOrNone(`
+          SELECT player.id AS player_id, elo_change.created_at, elo_change.elo_change
+          FROM PLAYER
+          JOIN LATERAL (
+            SELECT CASE
+              WHEN game.winner_id=player.id THEN game.winner_elo_change
+              WHEN game.loser_id=player.id THEN game.loser_elo_change
+            END as elo_change,
+            created_at
+            FROM game
+            WHERE (game.winner_id=player.id OR game.loser_id=player.id)
+              AND game.league_id=$1
+              AND created_at > CURRENT_DATE - INTERVAL '14 days'
+            ORDER BY game.created_at DESC
+          ) elo_change ON true
+          WHERE player.is_active=true
+          ORDER BY player.id
+        `, [league.id]),
+        db.manyOrNone('SELECT id, name, short_name FROM league WHERE is_active=true ORDER BY name ASC')
+      ])
+    })
+    .then((data) => {
+      const league = data[0]
+      const players = data[1]
+      const lastTengames = data[2]
+      const activeLeagues = data[3].map((league) => {
+        return {
+          name: league.name,
+          short_name: league.short_name,
+          isCurrentLeague: league.short_name === req.params.league_short_name
+        }
+      })
+
+      const gamesByPlayerId = {}
+      lastTengames.forEach((game) => {
+        gamesByPlayerId[game.player_id] = gamesByPlayerId[game.player_id] || []
+        gamesByPlayerId[game.player_id].push({
+          created_at: game.created_at,
+          elo_change: game.elo_change
+        })
+      })
+
+      const startDate = new Date()
+      const endDate = new Date(startDate.getTime())
+      endDate.setDate(endDate.getDate() - 14)
+
+      // Map of player_id -> [{player_name, elo_rating, date}]
+      const playerEloSeries = {}
+      players.forEach((player) => {
+        // Initialize array
+        // Current elo ranking is from last game to current time
+        playerEloSeries[player.id] = [
+          {
+            player_name: player.name,
+            color: player.color,
+            elo_rating: player.elo_rating,
+            date: startDate
+          }
+        ]
+        // If they have a last game then use this for next interval at current rank
+        const lastGames = gamesByPlayerId[player.id]
+        if (lastGames) {
+          playerEloSeries[player.id].push(
+            {
+              player_name: player.name,
+              color: player.color,
+              elo_rating: player.elo_rating,
+              date: gamesByPlayerId[player.id][0].created_at
+            }
+          )
+          // Iterate over last games
+          for (var i = 0; i < lastGames.length; i++) {
+            playerEloSeries[player.id].push(
+              {
+                player_name: player.name,
+                color: player.color,
+                elo_rating: playerEloSeries[player.id][playerEloSeries[player.id].length - 1].elo_rating - lastGames[i].elo_change,
+                date: lastGames[i + 1] ? lastGames[i + 1].created_at : endDate
+              }
+            )
+          }
+        } else {
+          // Append an end value
+          playerEloSeries[player.id].push(
+            {
+              player_name: player.name,
+              color: player.color,
+              elo_rating: playerEloSeries[player.id][playerEloSeries[player.id].length - 1].elo_rating,
+              date: endDate
+            }
+          )
+        }
+        // Reverse so time goes forward
+        playerEloSeries[player.id].reverse()
+      })
+
+      const gaugeData = []
+      for (var i = 0; i < players.length && i < 4; i++) {
+        gaugeData[i] = {
+          name: players[i].name,
+          color: players[i].color,
+          title: 'Games Won',
+          value: players[i].games_won / players[i].total_games
+        }
+      }
+
+      res.render('rankings', {
+        league: league,
+        players: players,
+        playerEloSeries: JSON.stringify(playerEloSeries),
+        currentLeague: req.params.league_short_name,
+        currentPage: '/rankings',
+        activeLeagues: activeLeagues,
+        gaugeData: JSON.stringify(gaugeData)
+      })
+    })
+    .catch((err) => {
+      res.render('error', {
+        error: err
+      })
+      console.error(err)
+    })
+  })
+}
